+++ rate.go
@@ -25,7 +25,8 @@ type Rate struct {
 	DeliveryDays           int        `json:"delivery_days,omitempty"`
 	DeliveryDate           *time.Time `json:"delivery_date,omitempty"`
 	DeliveryDateGuaranteed bool       `json:"delivery_date_guaranteed,omitempty"`
-	EstDeliveryDays        int        `json:"est_delivery_days,omitempty"`
+	EstDeliveryDays        int        `json:"est_delivery_dats,omitempty"`
+	BillingType            string     `json:"billing_type,omitempty"`
 }
 
 // A SmartRate contains information on shipping cost and delivery time in addition to time-in-transit details.