+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Next Release
 
-- Adds `PaymentMethod.ListPaymentMethods()`, `CreditCard.FundCreditCard()`, and `CreditCard.DeleteCreditCard()` functions
+- Adds PaymentMethod file with `ListPaymentMethods` method
+- Adds CreditCard file with `FundCreditCard`, and `DeleteCreditCard` methods
 
 ## v2.3.0 (2022-04-13)