+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Adds PaymentMethod file with `ListPaymentMethods` method
 - Adds CreditCard file with `FundCreditCard`, and `DeleteCreditCard` methods
 - Adds `BillingType` in CarrierAccount and Rate structs
+- Adds OS details to the User-Agent header
 
 ## v2.3.0 (2022-04-13)